+++ errors/collection.go
@@ -1,0 +1,6 @@
+package errors
+
+// ErrorCollection holds a list of errors.
+type ErrorCollection interface {
+	Errors() []error
+}